Do accept links to files w/o s3 prefix
lambdas.build_stac.utils.stac.gen_asset ignores a data asset if its link does not have an s3 prefix. We have valid data URLs for ESACCI biomass without s3 prefixes, for example, so I suggest to remove that conditional. In this example again, with that conditional we end up with items with missing data assets.